Move gwa status logic to kubeapi
